Revert "Add Events view"
