--- src/components/Events/index.js
@@ -1,5 +1,0 @@
-export { default as EventsTable } from './EventsTable'
-export * from './EventStatus'
-export * from './EventFilters'
-export { default as EventFilters } from './EventFilters'
-export { default as EventSort } from './EventSort'
