--- src/components/Events/style.css
@@ -1,3 +1,0 @@
-:global(#page-router-render-component) .container  {
-    padding-top: 0;
-}
