+++ src/components/VmDetails/cards/OverviewCard/style.css
@@ -2,31 +2,6 @@
  * Styles for the Overview Card
  */
 
-.bold {
-   font-weight: var(--pf-global--FontWeight--bold);
- }
-
- .fullWidth {
-  width: 100%;
-}
-
-.event {
-  display: flex;
-  /* event text and time label have different height */
-  align-items: center;
-}
-
-.eventText {
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  overflow: hidden;
-}
-
-
-.event :global(.pf-c-label) {
-  margin-right: 5px;
-}
-
 .operating-system-label {
   position: absolute;
   top: 0;
@@ -36,9 +11,7 @@
 
 .vm-info {
   padding: 15px;
-  flex-grow: 1;
-  display: flex;
-  flex-direction: column;
+  flex-grow: 1
 }
 
 .vm-name {
@@ -84,13 +57,13 @@
 }
 
 .os-icon {
-    padding: 15px;
+  padding: 15px;
 }
 
 .container {
   display: flex;
   overflow: visible;
-  flex-flow: row nowrap;
+  flex-flow: row wrap;
 }
 
 .pool-vm-label {
