+++ src/constants/index.js
@@ -5,13 +5,11 @@ export const ACTION_IN_PROGRESS_START = 'ACTION_IN_PROGRESS_START'
 export const ACTION_IN_PROGRESS_STOP = 'ACTION_IN_PROGRESS_STOP'
 export const ADD_ACTIVE_REQUEST = 'ADD_ACTIVE_REQUEST'
 export const ADD_DISK_REMOVAL_PENDING_TASK = 'ADD_DISK_REMOVAL_PENDING_TASK'
-export const ADD_LAST_VM_EVENTS = 'ADD_LAST_VM_EVENTS'
 export const ADD_NETWORKS_TO_VNIC_PROFILES = 'ADD_NETWORKS_TO_VNIC_PROFILES'
 export const ADD_SNAPSHOT_ADD_PENDING_TASK = 'ADD_SNAPSHOT_ADD_PENDING_TASK'
 export const ADD_SNAPSHOT_REMOVAL_PENDING_TASK = 'ADD_SNAPSHOT_REMOVAL_PENDING_TASK'
 export const ADD_SNAPSHOT_RESTORE_PENDING_TASK = 'ADD_SNAPSHOT_RESTORE_PENDING_TASK'
 export const ADD_VM_NIC = 'ADD_VM_NIC'
-export const ADD_VM_EVENTS = 'ADD_VM_EVENTS'
 export const ADD_USER_MESSAGE = 'ADD_USER_MESSAGE'
 export const APP_CONFIGURED = 'APP_CONFIGURED'
 export const AUTO_ACKNOWLEDGE = 'AUTO_ACKNOWLEDGE'
@@ -46,7 +44,6 @@ export const GET_POOL = 'GET_POOL'
 export const GET_POOLS = 'GET_POOLS'
 export const GET_VM = 'GET_VM'
 export const GET_VM_CDROM = 'GET_VM_CDROM'
-export const GET_VM_EVENTS = 'GET_VM_EVENTS'
 export const GET_VMS = 'GET_VMS'
 export const LOAD_USER_OPTIONS: 'LOAD_USER_OPTIONS' = 'LOAD_USER_OPTIONS'
 export const LOGIN = 'LOGIN'
@@ -70,7 +67,6 @@ export const REMOVE_VM = 'REMOVE_VM'
 export const RESTART_VM = 'RESTART_VM'
 export const SAVE_CONSOLE_OPTIONS = 'SAVE_CONSOLE_OPTIONS'
 export const SAVE_FILTERS = 'SAVE_FILTERS'
-export const SAVE_EVENT_FILTERS = 'SAVE_EVENT_FILTERS'
 export const SAVE_GLOBAL_OPTIONS: 'SAVE_GLOBAL_OPTIONS' = 'SAVE_GLOBAL_OPTIONS'
 export const SAVE_SSH_KEY = 'SAVE_SSH_KEY'
 export const SET_ADMINISTRATOR = 'SET_ADMINISTRATOR'
@@ -85,7 +81,6 @@ export const SET_CPU_TOPOLOGY_OPTIONS = 'SET_CPU_TOPOLOGY_OPTIONS'
 export const SET_CURRENT_PAGE = 'SET_CURRENT_PAGE'
 export const SET_DATA_CENTERS = 'SET_DATA_CENTERS'
 export const SET_DEFAULT_TIMEZONE = 'SET_DEFAULT_TIMEZONE'
-export const SET_EVENT_SORT = 'SET_EVENT_SORT'
 export const SET_FILTERS = 'SET_FILTERS'
 export const SET_HOSTS = 'SET_HOSTS'
 export const SET_OPERATING_SYSTEMS = 'SET_OPERATING_SYSTEMS'
