+++ src/constants/pages.js
@@ -5,4 +5,3 @@ export const LIST_PAGE_TYPE = 'listPage'
 export const NO_REFRESH_TYPE = 'noRefreshPage'
 export const EMPTY_VNIC_PROFILE_ID = ''
 export const SETTINGS_PAGE_TYPE = 'settingsPage'
-export const EVENTS_PAGE_TYPE = 'eventsPage'
