+++ src/intl/messages.js
@@ -80,7 +80,6 @@ export const messages: { [messageId: string]: MessageType } = {
   clear: 'Clear',
   clearAll: 'Clear all',
   clearAllFilters: 'Clear All Filters',
-  clearAllFiltersAndTryAgain: 'Clear all filters and try again',
   clearMessages: 'Clear Messages',
   clearAutoconnectVmNotAvailable: 'VM chosen for automatic console connection is no longer available. The setting will be cleared.',
   clickForHelp: 'Click for help',
@@ -190,7 +189,6 @@ export const messages: { [messageId: string]: MessageType } = {
   customScript: 'Custom script',
   dataCenter: { message: 'Data Center', description: 'Label for the VM\'s data center' },
   dataCenterChangesWithCluster: 'The data center cannot be edited from here. Please contact your administrator if you would like to edit the data center.',
-  date: 'Date',
   defaultButton: 'Default',
   defaultOption: {
     message: '(Default)',
@@ -268,10 +266,6 @@ export const messages: { [messageId: string]: MessageType } = {
   enum_DiskInterface_ide: { message: 'IDE', description: 'IDE controller VM disk attachment interface' },
   enum_DiskInterface_virtio: { message: 'VirtIO', description: 'virtio controller VM disk attachment interface' },
   enum_DiskInterface_virtio_scsi: { message: 'VirtIO-SCSI', description: 'virtio SCSI controller VM disk attachment interface' },
-  enum_EventSeverity_error: 'Error',
-  enum_EventSeverity_warning: 'Warning',
-  enum_EventSeverity_normal: 'Information',
-  enum_EventSeverity_unknown: 'Unknown',
   enum_NicInterface_e1000: {
     message: 'e1000',
     description: 'Display name of a NIC that provides an E1000 based interface to the VM',
@@ -358,9 +352,6 @@ export const messages: { [messageId: string]: MessageType } = {
   },
   error: 'Error',
   errorWhileCreatingNewDisk: 'Error while creating new disk:',
-  events: 'Events',
-  eventsFilterTypePlaceholderMessage: 'Filter by Message',
-  eventsFilterTypePlaceholderSeverity: 'Filter by Severity',
   every30Seconds: 'Every 30 seconds',
   everyMinute: 'Every minute',
   every2Minute: 'Every 2 minutes',
@@ -423,13 +414,11 @@ export const messages: { [messageId: string]: MessageType } = {
   ifVmIsRunningClickToAccessItsGraphicsConsole: 'If the virtual machine is running, click the protocol name to access its Graphical Console.',
   info: 'Information',
   inPreview: 'In Preview',
-  invalidDateFormat: 'Invalid date format. The supported format is {format}.',
   ieNotSupported: 'Internet Explorer is not a supported browser.',
   ipAddress: { message: 'IP Address', description: 'Label for IP addresses reported by VM guest agent' },
   isPersistMemorySnapshot: 'Content of the memory of the virtual machine is included in the snapshot.',
   itemDoesntExistOrDontHavePermissions: 'The item doesn\'t exist or you do not have the permissions to view it.',
   language: 'Language',
-  lastEvents: 'Last Events',
   less: {
     message: 'less',
     description: 'more/less pair used to control collapsible long listing',
@@ -458,7 +447,6 @@ export const messages: { [messageId: string]: MessageType } = {
   memory: 'Memory',
   memoryIncluded: '(State included)',
   messages: 'Messages',
-  message: 'Message',
   more: {
     message: 'more',
     description: 'more/less pair used to control collapsible long listing',
@@ -510,7 +498,6 @@ export const messages: { [messageId: string]: MessageType } = {
   noClustersAvailable: 'No Clusters available',
   noDisks: 'no disks',
   noError: 'No error',
-  noEventsFound: 'no events',
   noMessages: 'There are no notifications to display.',
   noneItem: '[None]',
   noNetwork: 'No network',
@@ -635,7 +622,6 @@ export const messages: { [messageId: string]: MessageType } = {
     message: 'Your session is about to timeout due to inactivity.',
     description: 'Primary message for SessionTimeout modal component',
   },
-  severity: 'Severity',
   shutdown: 'Shutdown',
   shutdownStatelessPoolVm: 'This virtual machine belongs to {poolName} and is stateless so any data that is currently attached to the virtual machine will be lost if it is shutdown. The virtual machine will be returned to {poolName} if shutdown.',
   shutdownVm: 'Shutdown the VM',
@@ -697,7 +683,6 @@ export const messages: { [messageId: string]: MessageType } = {
   timezone: 'Timezone',
   thisOperationCantBeUndone: 'This operation cannot be undone.',
   threadsPerCores: 'Threads per Core',
-  toggleDatePicker: 'Toggle date picker',
   totalCountOfVirtualProcessorsVmWillBeEquippedWith: 'Total count of virtual processors the virtual machine will be equipped with.',
   totalSocketsCpuTooltipMessage: '{number} virtual sockets',
   totalCoresCpuTooltipMessage: '{number} cores per socket',
@@ -784,7 +769,6 @@ export const messages: { [messageId: string]: MessageType } = {
   username: 'Username',
   usingRemoteViewer: 'Using a remote viewer relies on a downloaded .vv file.',
   vcpuTopology: 'VCPU Topology',
-  viewAll: 'View All',
   viewAllVirtualMachines: 'View All Virtual Machines',
   virtualMachines: 'Virtual Machines',
   virtualSockets: 'Virtual Sockets',
