+++ src/intl/translated-messages.json
@@ -45,7 +45,6 @@
     "empty": "Prázdné",
     "enterVmDescription": "Zadejte popis virtuálního stroje (volitelné)",
     "enterVmName": "Zadejte název virtuálního stroje",
-    "enum_EventSeverity_unknown": "Neznámý",
     "enum_VmStatus_down": "Vyp",
     "enum_VmStatus_image_locked": "Obraz disku uzamčen",
     "enum_VmStatus_migrating": "Migruje",
@@ -354,7 +353,6 @@
     "enum_DiskInterface_ide": "IDE",
     "enum_DiskInterface_virtio": "VirtIO",
     "enum_DiskInterface_virtio_scsi": "VirtIO-SCSI",
-    "enum_EventSeverity_unknown": "Unbekannt",
     "enum_NicInterface_e1000": "e1000",
     "enum_NicInterface_rtl8139": "rtl8139",
     "enum_NicInterface_virtio": "VirtIO",
@@ -889,7 +887,6 @@
     "enum_DiskInterface_ide": "IDE",
     "enum_DiskInterface_virtio": "VirtIO",
     "enum_DiskInterface_virtio_scsi": "VirtIO-SCSI",
-    "enum_EventSeverity_unknown": "Desconocido",
     "enum_NicInterface_e1000": "e1000",
     "enum_NicInterface_rtl8139": "rtl8139",
     "enum_NicInterface_virtio": "VirtIO",
@@ -1424,7 +1421,6 @@
     "enum_DiskInterface_ide": "IDE",
     "enum_DiskInterface_virtio": "VirtIO",
     "enum_DiskInterface_virtio_scsi": "VirtIO-SCSI",
-    "enum_EventSeverity_unknown": "Inconnue",
     "enum_NicInterface_e1000": "e1000",
     "enum_NicInterface_rtl8139": "rtl8139",
     "enum_NicInterface_virtio": "VirtIO",
@@ -2046,7 +2042,6 @@
     "enum_DiskInterface_ide": "IDE",
     "enum_DiskInterface_virtio": "VirtIO",
     "enum_DiskInterface_virtio_scsi": "VirtIO-SCSI",
-    "enum_EventSeverity_unknown": "Unknown",
     "enum_NicInterface_e1000": "e1000",
     "enum_NicInterface_rtl8139": "rtl8139",
     "enum_NicInterface_virtio": "VirtIO",
@@ -2435,7 +2430,6 @@
     "clear": "გასუფთავება",
     "clearAll": "ყველაფრის გასუფთავება",
     "clearAllFilters": "ყველა ფილტრის გასუფთავება",
-    "clearAllFiltersAndTryAgain": "გაასუფთავეთ ყველა ფილტრი და თავიდან სცადეთ",
     "clearAutoconnectVmNotAvailable": "კონსოლის ავტომატური დაკავშირებისთვის არჩეული ვმ-ი ხელმიუწვდომელია. პარამეტრი წაიშლება.",
     "clearMessages": "შეტყობინებების გასუფთავება",
     "clickForHelp": "დახმარებისთვის დააწკაპუნეთ",
@@ -2527,7 +2521,6 @@
     "customScript": "ხელით მითითებული სკრიპტი",
     "dataCenter": "მდც",
     "dataCenterChangesWithCluster": "მონაცემთა დამუშავების ცენტრის აქ ჩასწორება შეუძლებელია. მის ჩასასწორებლად დაუკავშირდით სისტემურ ადმინისტრატორს.",
-    "date": "თარიღი",
     "defaultButton": "ნაგულიხმები",
     "defaultOption": "(ნაგულისხმები)",
     "definedMemory": "აღწერილი მეხსიერება",
@@ -2596,10 +2589,6 @@
     "enum_DiskInterface_ide": "IDE",
     "enum_DiskInterface_virtio": "VirtIO",
     "enum_DiskInterface_virtio_scsi": "VirtIO-SCSI",
-    "enum_EventSeverity_error": "შეცდომა",
-    "enum_EventSeverity_normal": "ინფორმაცია",
-    "enum_EventSeverity_unknown": "უცნობი",
-    "enum_EventSeverity_warning": "გაფრთხილება",
     "enum_NicInterface_e1000": "e1000",
     "enum_NicInterface_e1000e": "e1000e",
     "enum_NicInterface_rtl8139": "rtl8139",
@@ -2623,9 +2612,6 @@
     "enum_VmStatus_wait_for_launch": "გაშვების მოლოდინი",
     "error": "შეცდომა",
     "errorWhileCreatingNewDisk": "ახალი დისკის შექმნის შეცდომა:",
-    "events": "მოვლენა",
-    "eventsFilterTypePlaceholderMessage": "შეტყობინებით გაფილტვრა",
-    "eventsFilterTypePlaceholderSeverity": "სიმწვავის მიხედვით გაშიფვრა",
     "every2Minute": "ყოველ 2 წუთში",
     "every30Seconds": "ყოველ 30 წამში",
     "every5Minute": "ყოველ 5 წუთში",
@@ -2678,12 +2664,10 @@
     "ifVmIsRunningClickToAccessItsGraphicsConsole": "თუ ვირტუალური მანქანა გაშვებულია, კონსოლის გასაშვებად დააწკაპუნეთ პროტოკოლის შესაბამის სახელს.",
     "inPreview": "გადახედვა",
     "info": "ინფორმაცია",
-    "invalidDateFormat": "თარიღის არასწორი ფორმატი. მხარდაჭერილი ფორმატია {format}.",
     "ipAddress": "IP მისამართი",
     "isPersistMemorySnapshot": "სწრაფი ასლი ვირტუალური მანქანის მეხსიერების შემცველობასაც შეიცავს.",
     "itemDoesntExistOrDontHavePermissions": "ჩანაწერი არ არსებობს ან არ გაქვთ მისი ნახვის უფლება.",
     "language": "ენა",
-    "lastEvents": "ბოლო მოვლენები",
     "less": "ნაკლები",
     "loadingTripleDot": "ჩატვირთვა ...",
     "locked": "დაბლოკილია",
@@ -2702,7 +2686,6 @@
     "maxNumberOfVms": "ამ პულიდან თქვენი მოხმარებისათვის მაქსიმუმ {numberOfVms} შეიძლება გამოიყოს.",
     "memory": "მეხსიერება",
     "memoryIncluded": "(შეიცავს მდგომარეობასაც)",
-    "message": "შეტყობინება",
     "messages": "შეტყობინებები",
     "more": "მეტი",
     "name": "სახელი",
@@ -2740,7 +2723,6 @@
     "noClustersAvailable": "კლასტერები ხელმიუწვდომელია",
     "noDisks": "დისკების გარეშე",
     "noError": "შეცდომის გარეშე",
-    "noEventsFound": "მოვლენების გარეშე",
     "noMessages": "საჩვენებელი შეტყობინებების გარეშე.",
     "noNetwork": "ქსელის გარეშე",
     "noNics": "ქსელის ინტერფეისების გარეშე",
@@ -2823,7 +2805,6 @@
     "serialConsole": "სერიული კონსოლი",
     "serialConsoleOptions": "სერიული კონსოლის მორგება",
     "sessionExpired": "თქვენი სესია არააქტიურობის გამო მალე დასრულდება.",
-    "severity": "სიმწვავე",
     "shutdown": "გამორთვა",
     "shutdownStatelessPoolVm": "ვირტუალური მანქანა ეკუთვნის პულს ( {poolName} ) და არის უნიადაგო, ამიტომ ყველა მონაცემი, რომელიც მიმაგრებულია, წაიშლება მისი გათიშვისთანავე. ვირტუალური მანქანა დაბრუნდება {poolName} -ში ",
     "shutdownVm": "ვმ-ის გამორთვა",
@@ -2873,7 +2854,6 @@
     "thisOperationCantBeUndone": "ამ ოპერაციის დაბრუნება შეუძლებელია.",
     "threadsPerCores": "ნაკადები თითოეული ბირთვისთვის",
     "timezone": "დროის სარტყელი",
-    "toggleDatePicker": "თარიღის ამრჩევის გადართვა",
     "totalCoresCpuTooltipMessage": "{number} ბირთვი თითოეულ სოკეტზე",
     "totalCountOfVirtualProcessorsVmWillBeEquippedWith": "ვირტუალური მანქანისთვის განკუთვნილი ვირტუალური პროცესორების რაოდენობა.",
     "totalMemoryVmWillBeEquippedWith": "ვირტუალური მანქანისთვის განკუთვნილი მეხსიერების რაოდენობა.",
@@ -2927,7 +2907,6 @@
     "utilizationNoHistoricData": "ისტორიის მონაცემები ხელმიუწვდომელია",
     "utilizationNoNetStats": "ამ ვმ-სთვის ქსელის გამოყენება ხელმიუწვდომელია",
     "vcpuTopology": "VCPU-ის ტოპოლოგია",
-    "viewAll": "ყველას ნახვა",
     "viewAllVirtualMachines": "ყველა ვირტუალური მანქანის ნახვა",
     "virtualMachines": "ვირტუალური მანქანები",
     "virtualSockets": "ვირტუალური სოკეტები",
@@ -3169,7 +3148,6 @@
     "enum_DiskInterface_ide": "IDE",
     "enum_DiskInterface_virtio": "VirtIO",
     "enum_DiskInterface_virtio_scsi": "VirtIO-SCSI",
-    "enum_EventSeverity_unknown": "알 수 없음 ",
     "enum_NicInterface_e1000": "e1000",
     "enum_NicInterface_rtl8139": "rtl8139",
     "enum_NicInterface_virtio": "VirtIO",
@@ -3704,7 +3682,6 @@
     "enum_DiskInterface_ide": "IDE",
     "enum_DiskInterface_virtio": "VirtIO",
     "enum_DiskInterface_virtio_scsi": "VirtIO-SCSI",
-    "enum_EventSeverity_unknown": "Desconhecido",
     "enum_NicInterface_e1000": "e1000",
     "enum_NicInterface_rtl8139": "rtl8139",
     "enum_NicInterface_virtio": "VirtIO",
@@ -4239,7 +4216,6 @@
     "enum_DiskInterface_ide": "IDE",
     "enum_DiskInterface_virtio": "VirtIO",
     "enum_DiskInterface_virtio_scsi": "VirtIO-SCSI",
-    "enum_EventSeverity_unknown": "未知",
     "enum_NicInterface_e1000": "e1000",
     "enum_NicInterface_rtl8139": "rtl8139",
     "enum_NicInterface_virtio": "VirtIO",
