+++ src/routes.js
@@ -9,14 +9,12 @@ import {
   VmDetailToolbar,
   VmsListToolbar,
   SettingsToolbar,
-  EventsToolbar,
 } from './components/Toolbar'
 import {
   VmDetailsPage,
   VmsListPage,
   GlobalSettingsPage,
   VmConsolePage,
-  VmEventsPage,
 } from './components/Pages'
 
 import {
@@ -25,7 +23,6 @@ import {
   CONSOLE_PAGE_TYPE,
   NO_REFRESH_TYPE,
   SETTINGS_PAGE_TYPE,
-  EVENTS_PAGE_TYPE,
 } from '_/constants'
 
 /**
@@ -67,15 +64,6 @@ export default function getRoutes () {
                 toolbars: (match) => null,
                 isToolbarFullWidth: true,
                 type: CONSOLE_PAGE_TYPE,
-              },
-              {
-                path: '/vm/:id/events',
-                title: ({ msg }) => msg.events(),
-                component: VmEventsPage,
-                closeable: true,
-                toolbars: (match) => <EventsToolbar match={match}/>,
-                isToolbarFullWidth: true,
-                type: EVENTS_PAGE_TYPE,
               },
             ],
           },
